refactor: split the last four oversized view files
ios/…/TransactionsView.swift  1048 -> 463
  TransactionForms.swift    the New/Edit Transaction and Transfer sheets
  CategoryBreakdown.swift   the Top-categories donut, chips and period picker

ios/…/DashboardView.swift    1009 -> 360
  DashboardDerived.swift    the visibility key and the rollups
  DashboardCharts.swift     the net-worth area and split charts
  DashboardRows.swift       account, transaction, breakdown and runway rows

frontend/…/Portfolio.tsx     1057 -> 878
  PortfolioPanels.tsx       move cash, edit asset, record price, rename

Both Swift files were already a main view followed by standalone structs,
so those moves are relocation with no call-site changes.

The Portfolio panels keep the page's own prop names, so the markup moved
verbatim rather than being retyped. Each returns its own `&& (…)` guard
instead of being wrapped in one at the call site, which leaves the page's
JSX a flat list of panels rather than four more nested conditionals.

Portfolio.tsx is now under 1000. Every file named in this pass is.
